estiam_service/requirements.txt
asgiref==3.8.1
attrs==23.2.0
cfgv==3.4.0
distlib==0.3.8
Django==5.0.4
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
factory-boy==3.3.0
Faker==24.9.0
filelock==3.13.4
identify==2.5.35
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
nodeenv==1.8.0
packaging==24.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.7.0
PyJWT==2.8.0
pytest==8.1.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.5.0
uritemplate==4.1.1
virtualenv==20.25.2
